moving ITs to ws.it package to keep consistency with other GBIF projects
