+++ occurrence-integration-tests/pom.xml
@@ -40,6 +40,10 @@
           <groupId>org.springframework.boot</groupId>
           <artifactId>spring-boot-starter-tomcat</artifactId>
         </exclusion>
+        <exclusion>
+          <groupId>org.elasticsearch.client</groupId>
+          <artifactId>elasticsearch-rest-client</artifactId>
+        </exclusion>
       </exclusions>
     </dependency>
     <dependency>
@@ -78,15 +82,15 @@
       </exclusions>
     </dependency>
     <dependency>
-      <groupId> org.elasticsearch.client</groupId>
+      <groupId>org.elasticsearch.client</groupId>
       <artifactId>elasticsearch-rest-client</artifactId>
     </dependency>
     <dependency>
-      <groupId> org.elasticsearch</groupId>
+      <groupId>org.elasticsearch</groupId>
       <artifactId>elasticsearch-core</artifactId>
     </dependency>
     <dependency>
-      <groupId> org.elasticsearch</groupId>
+      <groupId>org.elasticsearch</groupId>
       <artifactId>elasticsearch</artifactId>
     </dependency>
     <dependency>
@@ -115,6 +119,12 @@
     <dependency>
       <groupId>org.gbif.pipelines</groupId>
       <artifactId>ingest-gbif-java</artifactId>
+      <exclusions>
+        <exclusion>
+          <groupId>org.elasticsearch.client</groupId>
+          <artifactId>elasticsearch-rest-high-level-client</artifactId>
+        </exclusion>
+      </exclusions>
     </dependency>
     <dependency>
       <groupId>org.gbif.pipelines</groupId>
