+++ occurrence-integration-tests/src/test/java/org/gbif/occurrence/ws/it/OccurrenceSearchResourceIT.java
@@ -1,4 +1,4 @@
-package org.gbif.occurrence.it.ws;
+package org.gbif.occurrence.ws.it;
 
 import org.gbif.api.model.common.search.Facet;
 import org.gbif.api.model.common.search.SearchResponse;
